Fixing My Own Errors
Bot no doesn't try to read the .status off a dictionary, that's now done properly. The 'content_type=None' silences the mimetype error when it appears and allows the bot to send 'Unknown IGN' in the given channel.

This fix is for all effected commands EXCEPT for ',blacklist' and ',dnkladd', which as of the current moment still get the MimeType error, even with 'content_type=None' implemented.