Gexp Requirement Checking
Most of the general logic for roll_giveaway has been completed, and the bot now checks for gexp requirements.

Fixed the incorrect input of gexp requirements to the json and syntax and stylistic problems.

The time of a giveaway ending is now calculated when the giveaway is created, not when the time is entered.