fix prefix help
This PR fixes the help command using a prefix instead of always using the slash command version. the only downfall is that it shows the "Hidden cog" and a few commands. 

There is a bug that shows "Help None" at the bottom of the embed but it does noy seem to error upon using the command anymore.